+++ src/i_winmusic.c
@@ -77,18 +77,6 @@ typedef struct
     int                 absolute_time;
 } win_midi_track_t;
 
-static const int volume_correction[] =
-{
-      0,   4,   7,  11,  13,  14,  16,  18,  21,  22,  23,  24,  24,  24,  25,  25,
-     25,  26,  26,  27,  27,  27,  28,  28,  29,  29,  29,  30,  30,  31,  31,  32,
-     32,  32,  33,  33,  34,  34,  35,  35,  36,  37,  37,  38,  38,  39,  39,  40,
-     40,  41,  42,  42,  43,  43,  44,  45,  45,  46,  47,  47,  48,  49,  49,  50,
-     51,  52,  52,  53,  54,  55,  56,  56,  57,  58,  59,  60,  61,  62,  62,  63,
-     64,  65,  66,  67,  68,  69,  70,  71,  72,  73,  74,  75,  77,  78,  79,  80,
-     81,  82,  84,  85,  86,  87,  89,  90,  91,  92,  94,  95,  96,  98,  99, 101,
-    102, 104, 105, 107, 108, 110, 112, 113, 115, 117, 118, 120, 122, 123, 125, 127
-};
-
 static float    volume_factor = 1.0f;
 
 // Save the last volume for each MIDI channel.
@@ -138,7 +126,7 @@ static void FillBuffer(void)
 
             channel_volume[MIDIEVENT_CHANNEL(event->dwEvent)] = volume;
             event->dwEvent = ((event->dwEvent & 0xFF00FFFF)
-                | ((volume_correction[(int)((float)volume * volume_factor)] & 0x7F) << 16));
+                | (((int)((float)volume * volume_factor) & 0x7F) << 16));
         }
 
         song.position++;
@@ -297,6 +285,14 @@ static void MIDItoStream(midi_file_t *file)
         free(tracks);
 }
 
+static void UpdateVolume(void)
+{
+    // Send MIDI controller events to adjust the volume.
+    for (int i = 0; i < MIDI_CHANNELS_PER_TRACK; i++)
+        midiOutShortMsg((HMIDIOUT)hMidiStream, (MIDI_EVENT_CONTROLLER | i | (MIDI_CONTROLLER_MAIN_VOLUME << 8)
+            | ((int)((float)channel_volume[i] * volume_factor) << 16)));
+}
+
 dboolean I_Windows_InitMusic(void)
 {
     UINT    MidiDevice = MIDI_MAPPER;
@@ -324,11 +320,7 @@ dboolean I_Windows_InitMusic(void)
 void I_Windows_SetMusicVolume(int volume)
 {
     volume_factor = (float)volume / MIX_MAX_VOLUME;
-
-    // Send MIDI controller events to adjust the volume.
-    for (int i = 0; i < MIDI_CHANNELS_PER_TRACK; i++)
-        midiOutShortMsg((HMIDIOUT)hMidiStream, (MIDI_EVENT_CONTROLLER | i | (MIDI_CONTROLLER_MAIN_VOLUME << 8)
-            | (volume_correction[(int)((float)channel_volume[i] * volume_factor)] << 16)));
+    UpdateVolume();
 }
 
 void I_Windows_StopSong(void)
@@ -369,6 +361,8 @@ void I_Windows_PlaySong(dboolean looping)
     SetThreadPriority(hPlayerThread, THREAD_PRIORITY_TIME_CRITICAL);
 
     midiStreamRestart(hMidiStream);
+
+    UpdateVolume();
 }
 
 void I_Windows_RegisterSong(void *data, int size)